requirements.txt
requests
beautifulsoup4
pymupdf
trafilatura
openai
supabase
python-dotenv
tiktoken
numpy
langchain
openai
chromadb
tiktoken
unstructured
python-dotenv
sentence-transformers

langchain
langchain-community 
langchain-openai